+++ test/group-by/e2e.spec.ts
@@ -6,7 +6,7 @@ import * as path from 'path'
 import { wait } from 'payload/shared'
 import { fileURLToPath } from 'url'
 
-import type { PayloadTestSDK } from '../helpers/shared/sdk/index.js'
+import type { PayloadTestSDK } from '../__helpers/shared/sdk/index.js'
 import type { Config, Post } from './payload-types.js'
 
 import {
@@ -15,21 +15,21 @@ import {
   initPageConsoleErrorCatch,
   saveDocAndAssert,
   selectTableRow,
-} from '../helpers/e2e/helpers.js'
-import { AdminUrlUtil } from '../helpers/shared/adminUrlUtil.js'
-import { sortColumn, toggleColumn } from '../helpers/e2e/columns/index.js'
-import { addListFilter } from '../helpers/e2e/filters/index.js'
-import { goToNextPage } from '../helpers/e2e/goToNextPage.js'
+} from '../__helpers/e2e/helpers.js'
+import { AdminUrlUtil } from '../__helpers/shared/adminUrlUtil.js'
+import { sortColumn, toggleColumn } from '../__helpers/e2e/columns/index.js'
+import { addListFilter } from '../__helpers/e2e/filters/index.js'
+import { goToNextPage } from '../__helpers/e2e/goToNextPage.js'
 import {
   addGroupBy,
   clearGroupBy,
   closeGroupBy,
   openGroupBy,
-} from '../helpers/e2e/groupBy/index.js'
-import { deletePreferences } from '../helpers/e2e/preferences.js'
-import { openNav } from '../helpers/e2e/toggleNav.js'
-import { initPayloadE2ENoConfig } from '../helpers/shared/initPayloadE2ENoConfig.js'
-import { reInitializeDB } from '../helpers/shared/clearAndSeed/reInitializeDB.js'
+} from '../__helpers/e2e/groupBy/index.js'
+import { deletePreferences } from '../__helpers/e2e/preferences.js'
+import { openNav } from '../__helpers/e2e/toggleNav.js'
+import { initPayloadE2ENoConfig } from '../__helpers/shared/initPayloadE2ENoConfig.js'
+import { reInitializeDB } from '../__helpers/shared/clearAndSeed/reInitializeDB.js'
 import { TEST_TIMEOUT_LONG } from '../playwright.config.js'
 import { postsSlug } from './collections/Posts/index.js'
