+++ test/joins/buildJoinAggregation.int.spec.ts
@@ -9,7 +9,7 @@ import { expect, it } from 'vitest'
 
 import { buildJoinAggregation } from '../../packages/db-mongodb/src/utilities/buildJoinAggregation.js'
 import { buildProjectionFromSelect } from '../../packages/db-mongodb/src/utilities/buildProjectionFromSelect.js'
-import { describe } from '../helpers/int/vitest.js'
+import { describe } from '../__helpers/int/vitest.js'
 
 describe(
   'buildJoinAggregation',
