+++ test/lexical/collections/LexicalJSXConverter/e2e.spec.ts
@@ -1,7 +1,7 @@
 import type { Locator, Page } from '@playwright/test'
 
 import { expect, test } from '@playwright/test'
-import { AdminUrlUtil } from '../../../helpers/shared/adminUrlUtil.js'
+import { AdminUrlUtil } from '../../../__helpers/shared/adminUrlUtil.js'
 import { reInitializeDB } from 'helpers/shared/clearAndSeed/reInitializeDB.js'
 import { lexicalJSXConverterSlug } from 'lexical/slugs.js'
 import path from 'path'
