+++ test/localization/e2e.spec.ts
@@ -10,7 +10,7 @@ import { upsertPreferences } from 'helpers/e2e/preferences.js'
 import { runAxeScan } from 'helpers/e2e/runAxeScan.js'
 import { openDocDrawer } from 'helpers/e2e/toggleDocDrawer.js'
 import { waitForAutoSaveToRunAndComplete } from 'helpers/e2e/waitForAutoSaveToRunAndComplete.js'
-import { RESTClient } from '../helpers/shared/rest.js'
+import { RESTClient } from '../__helpers/shared/rest.js'
 import path from 'path'
 import { formatAdminURL } from 'payload/shared'
 import { fileURLToPath } from 'url'
