+++ test/plugin-form-builder/int.spec.ts
@@ -9,7 +9,7 @@ import type { Form } from './payload-types.js'
 
 import { serializeLexical } from '../../packages/plugin-form-builder/src/utilities/lexical/serializeLexical.js'
 import { serializeSlate } from '../../packages/plugin-form-builder/src/utilities/slate/serializeSlate.js'
-import { initPayloadInt } from '../helpers/shared/initPayloadInt.js'
+import { initPayloadInt } from '../__helpers/shared/initPayloadInt.js'
 import { formsSlug, formSubmissionsSlug } from './shared.js'
 
 let payload: Payload
