+++ test/queues/postgres-logs.int.spec.ts
@@ -5,7 +5,7 @@ import path from 'path'
 import { fileURLToPath } from 'url'
 import { afterAll, beforeAll, describe, expect, it, vitest } from 'vitest'
 
-import { initPayloadInt } from '../helpers/shared/initPayloadInt.js'
+import { initPayloadInt } from '../__helpers/shared/initPayloadInt.js'
 import { withoutAutoRun } from './utilities.js'
 
 const filename = fileURLToPath(import.meta.url)
